Correction of `gradle_command` string value
Since you are already running the `cd_command` to go inside the project directory, there should not be `project_path` before "gradlew".
I was having `System cannot find the file specified` issue when I was using it as it was. So, I changed the library code to generate the APK.
BEFORE:
```
cd /path/to/project
/path/to/project/gradlew assembleRelease
```

AFTER:
```
cd /path/to/project
gradlew assembleRelease
```
Corrected!